requirements.txt
Flask==2.3.3
reportlab==4.0.5
yt-dlp==2025.12.8
python-dotenv==1.0.1
beautifulsoup4==4.12.2
requests==2.31.0